Change stacktrace format and optimize generating stacktraces.

```
$ benchstat old.log new.log
name                      old time/op    new time/op    delta
ErrWithStackAndUTCTime-2    21.9µs ± 7%    16.8µs ± 3%  -23.24%  (p=0.000 n=9+10)
ErrWithStack-2              20.8µs ± 3%    16.1µs ± 4%  -22.82%  (p=0.000 n=10+10)

name                      old alloc/op   new alloc/op   delta
ErrWithStackAndUTCTime-2    16.6kB ± 0%    16.5kB ± 0%   -0.55%  (p=0.000 n=10+10)
ErrWithStack-2              16.6kB ± 0%    16.5kB ± 0%   -0.57%  (p=0.000 n=10+10)

name                      old allocs/op  new allocs/op  delta
ErrWithStackAndUTCTime-2      5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
ErrWithStack-2                5.00 ± 0%      4.00 ± 0%  -20.00%  (p=0.000 n=10+10)
```